+++ multimedia-slides-library/index.js
@@ -3,9 +3,9 @@ import {Slide} from "./modules/Slide.js";
 import {RasterGraphicsSlide} from "./modules/RasterGraphicsSlide.js";
 import {VideoSlide} from "./modules/VideoSlide.js";
 import {SlideLibrary} from "./modules/SlideLibrary.js";
-import {SlideStorage} from "./modules/SlideStorage.js";
-import {UIManager} from "./modules/UIManager.js";
-import {ToolManager} from "./modules/ToolManager.js";
+import {SlideStorage} from "./modules/utils/SlideStorage.js";
+import {UIManager} from "./modules/utils/UIManager.js";
+import {ToolManager} from "./modules/utils/ToolManager.js";
 
 
 export {Slide};
