+++ multimedia-slides-library/modules/utils/SlideStorage.js
