+++ multimedia-slides-library/modules/utils/ToolManager.js
