+++ multimedia-slides-library/modules/utils/UIManager.js
