--- raster-slide-example.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Raster Slide Example</title>
-    <style>
-        #slide-container { width: 600px; height: 400px; border: 1px solid black; }
-        #tools-container { margin: 10px 0; }
-        #controls-container { margin: 10px 0; }
-        #playback-controls-container { margin: 10px 0; }
-    </style>
-</head>
-<body>
-<div id="slide-container"></div>
-<div id="tools-container"></div>
-<div id="controls-container">
-    <button id="start-recording-button">Начать запись</button>
-    <button id="stop-recording-button" disabled>Остановить запись</button>
-</div>
-<div id="playback-controls-container">
-    <h2>Записанные наборы команд:</h2>
-    <div id="playback-buttons-container"></div>
-</div>
-
-<!-- Ссылка на пример с видео -->
-<p><a href="video-slide-example.html">Перейти к примеру с видео</a></p>
-
-<script type="module" src="examples/RasterSlideUsageExample.js"></script>
-</body>
-</html>
